Rework scripts to compile and install everything without root.
Paths are specified in a separate config file.
By default sources are downloaded to ~/rigs-of-rods/source,
and binaries are installed to ~/rigs-of-rods/install.

Added "set -eu" to stop on errors or undefined variables.
Added "-c" to wget, to continue interrupted transfers.
Added "-o" to unzip, so it doesn't stop if files are already there.
Added "git pull" to update git repositories when the script is run again.